requirements.txt
# JiuCi AI伴侣 - 项目依赖包

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 异步支持
asyncio-mqtt==0.16.1

# 数据处理
pydantic==2.4.2
dataclasses-json==0.6.1

# 语音处理
websockets==11.0.3
volcengine-python-sdk==1.0.98

# 情感分析
transformers==4.35.0
torch==2.1.0

# 工具库
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# 日志和配置
loguru==0.7.2

# 开发工具
pytest==7.4.3
black==23.9.1
flake8==6.1.0

# 可选依赖（根据需要安装）
# openai==0.28.1  # 如果使用OpenAI API
# azure-cognitiveservices-speech==1.34.0  # 如果使用Azure语音服务